# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  avatarColor: string;
  createdAt: string;
}

export interface CreateUserInput {
  username: string;
  displayName: string;
}

export interface Contact {
  id: number;
  userId: number;
  contactId: number;
  contact: User;
  createdAt: string;
}

export interface AddContactInput {
  contactId: number;
}

export type ConversationType =
  (typeof ConversationType)[keyof typeof ConversationType];

export const ConversationType = {
  dm: "dm",
  group: "group",
} as const;

export type ChatMessageType =
  (typeof ChatMessageType)[keyof typeof ChatMessageType];

export const ChatMessageType = {
  text: "text",
  image: "image",
  file: "file",
  video: "video",
} as const;

export interface ChatMessage {
  id: number;
  conversationId: number;
  senderId: number;
  sender: User;
  content: string;
  type: ChatMessageType;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
  createdAt: string;
}

export interface Conversation {
  id: number;
  type: ConversationType;
  name?: string;
  members: User[];
  lastMessage?: ChatMessage;
  createdAt: string;
}

export type CreateConversationInputType =
  (typeof CreateConversationInputType)[keyof typeof CreateConversationInputType];

export const CreateConversationInputType = {
  dm: "dm",
  group: "group",
} as const;

export interface CreateConversationInput {
  type: CreateConversationInputType;
  name?: string;
  memberIds: number[];
}

export type SendChatMessageInputType =
  (typeof SendChatMessageInputType)[keyof typeof SendChatMessageInputType];

export const SendChatMessageInputType = {
  text: "text",
  image: "image",
  file: "file",
  video: "video",
} as const;

export interface SendChatMessageInput {
  senderId: number;
  content: string;
  type: SendChatMessageInputType;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type SearchUsersParams = {
  q: string;
};

export type ListConversationsParams = {
  userId: number;
};

export type ListMessagesParams = {
  before?: number;
};