# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create or get a user by username
 */
export const CreateOrGetUserBody = zod.object({
  username: zod.string(),
  displayName: zod.string(),
});

export const CreateOrGetUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  avatarColor: zod.string(),
  createdAt: zod.date(),
});

/**
 * @summary Search users by username
 */
export const SearchUsersQueryParams = zod.object({
  q: zod.coerce.string(),
});

export const SearchUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  displayName: zod.string(),
  avatarColor: zod.string(),
  createdAt: zod.date(),
});
export const SearchUsersResponse = zod.array(SearchUsersResponseItem);

/**
 * @summary Get user's contacts
 */
export const GetContactsParams = zod.object({
  userId: zod.coerce.number(),
});

export const GetContactsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  contactId: zod.number(),
  contact: zod.object({
    id: zod.number(),
    username: zod.string(),
    displayName: zod.string(),
    avatarColor: zod.string(),
    createdAt: zod.date(),
  }),
  createdAt: zod.date(),
});
export const GetContactsResponse = zod.array(GetContactsResponseItem);

/**
 * @summary Add a contact
 */
export const AddContactParams = zod.object({
  userId: zod.coerce.number(),
});

export const AddContactBody = zod.object({
  contactId: zod.number(),
});

/**
 * @summary Remove a contact
 */
export const RemoveContactParams = zod.object({
  userId: zod.coerce.number(),
  contactId: zod.coerce.number(),
});

/**
 * @summary List conversations for a user
 */
export const ListConversationsQueryParams = zod.object({
  userId: zod.coerce.number(),
});

export const ListConversationsResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["dm", "group"]),
  name: zod.string().optional(),
  members: zod.array(
    zod.object({
      id: zod.number(),
      username: zod.string(),
      displayName: zod.string(),
      avatarColor: zod.string(),
      createdAt: zod.date(),
    }),
  ),
  lastMessage: zod
    .object({
      id: zod.number(),
      conversationId: zod.number(),
      senderId: zod.number(),
      sender: zod.object({
        id: zod.number(),
        username: zod.string(),
        displayName: zod.string(),
        avatarColor: zod.string(),
        createdAt: zod.date(),
      }),
      content: zod.string(),
      type: zod.enum(["text", "image", "file", "video"]),
      fileUrl: zod.string().optional(),
      fileName: zod.string().optional(),
      fileType: zod.string().optional(),
      createdAt: zod.date(),
    })
    .optional(),
  createdAt: zod.date(),
});
export const ListConversationsResponse = zod.array(
  ListConversationsResponseItem,
);

/**
 * @summary Create a DM or group conversation
 */
export const CreateConversationBody = zod.object({
  type: zod.enum(["dm", "group"]),
  name: zod.string().optional(),
  memberIds: zod.array(zod.number()),
});

/**
 * @summary Get messages in a conversation
 */
export const ListMessagesParams = zod.object({
  conversationId: zod.coerce.number(),
});

export const ListMessagesQueryParams = zod.object({
  before: zod.coerce.number().optional(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  senderId: zod.number(),
  sender: zod.object({
    id: zod.number(),
    username: zod.string(),
    displayName: zod.string(),
    avatarColor: zod.string(),
    createdAt: zod.date(),
  }),
  content: zod.string(),
  type: zod.enum(["text", "image", "file", "video"]),
  fileUrl: zod.string().optional(),
  fileName: zod.string().optional(),
  fileType: zod.string().optional(),
  createdAt: zod.date(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a message in a conversation
 */
export const SendChatMessageParams = zod.object({
  conversationId: zod.coerce.number(),
});

export const SendChatMessageBody = zod.object({
  senderId: zod.number(),
  content: zod.string(),
  type: zod.enum(["text", "image", "file", "video"]),
  fileUrl: zod.string().optional(),
  fileName: zod.string().optional(),
  fileType: zod.string().optional(),
});

/**
 * @summary List all AI conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new AI conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get AI conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete an AI conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in an AI conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary Generate an image from a text prompt
 */
export const GenerateOpenaiImageBody = zod.object({
  prompt: zod.string(),
  size: zod.enum(["1024x1024", "512x512", "256x256"]).optional(),
});

export const GenerateOpenaiImageResponse = zod.object({
  b64_json: zod.string(),
});

/**
 * @summary Request a presigned URL for file upload
 */

export const RequestUploadUrlBody = zod.object({
  name: zod.string().min(1),
  size: zod.number().min(1),
  contentType: zod.string().min(1),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string().url(),
  objectPath: zod.string(),
  metadata: zod
    .object({
      name: zod.string().min(1),
      size: zod.number().min(1),
      contentType: zod.string().min(1),
    })
    .optional(),
});

/**
 * @summary Serve an object entity
 */
export const GetStorageObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});