# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageType } from "./chatMessageType";
import type { User } from "./user";

export interface ChatMessage {
  id: number;
  conversationId: number;
  senderId: number;
  sender: User;
  content: string;
  type: ChatMessageType;
  fileUrl?: string;
  fileName?: string;
  fileType?: string;
  createdAt: Date;
}