# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Contact {
  id: number;
  userId: number;
  contactId: number;
  contact: User;
  createdAt: Date;
}