# lib/api-zod/src/generated/types/createConversationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateConversationInputType } from "./createConversationInputType";

export interface CreateConversationInput {
  type: CreateConversationInputType;
  name?: string;
  memberIds: number[];
}