# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Private Chat App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserInput {
  username: string;
  displayName: string;
}